refactor(core): Move the update_workflow pipeline steps into their own module (no-changelog)
update-workflow.tool.ts is now 342 lines holding only the imports, the
tool factory and the handler; the ten pipeline steps move to a new
partial-update-steps.ts, named to pair with partial-update-schemas.ts.
dedupeNamesPreservingCase and countOperationsWithNoEffect stay private to
that module.

Rewriting the import block also cleared the 14 import-x/order warnings
the tool file had been carrying.

No behaviour change: all 121 existing tests pass untouched.
